Minor code style refactor
Use object property lookup instead of a loop.
Rename variables a bit.

Extract creating the HTML for the Messenger message to make the code
easier to read.

Fix storing the first message text in the cache.

Add a TODO comment to use MutationObserver for improved performance.

Use `queryElement` instead of `getElementsByTagName`.